adminstudentsp/backend/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
flask_cors
flask_migrate
python-dotenv
